# Optimize AspNet Core propagation extract