The request object already comes with a dictionary like object so all the loop allocations can be avoid with a simple wrapper. The measurement below is only for the ExtractPropagatedContext method.

|      Method | Categories |       Mean |    Error |   StdDev |     Median | Ratio | RatioSD |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------ |----------- |-----------:|---------:|---------:|-----------:|------:|--------:|-------:|------:|------:|----------:|
|        NoB3 |       NoB3 | 1,100.0 ns | 26.09 ns | 75.26 ns | 1,078.7 ns |  1.00 |    0.00 | 0.3548 |     - |     - |    1488 B |
|     NewNoB3 |       NoB3 |   427.4 ns |  8.45 ns | 12.90 ns |   424.1 ns |  0.39 |    0.03 | 0.0858 |     - |     - |     360 B |
|             |            |            |          |          |            |       |         |        |       |       |           |
|    B3Header |         B3 | 1,907.0 ns | 32.27 ns | 38.42 ns | 1,893.4 ns |  1.00 |    0.00 | 0.5226 |     - |     - |    2200 B |
| NewB3Header |         B3 |   998.1 ns | 19.55 ns | 24.72 ns |   995.1 ns |  0.52 |    0.02 | 0.1984 |     - |     - |     832 B |